requirements.txt
langchain-google-genai
google-generativeai
langchain_core
python-dotenv
streamlit
langchain_community
